feat: Breaks up message service into smaller pieces
