+++ src/services/messages/display/Components/index.ts
@@ -1,0 +1,16 @@
+export * from './cancelSoftwareUpgradeProposal';
+export * from './instantiateCodeProposal';
+export * from './msgAddMarkerRequest';
+export * from './msgGovVote';
+export * from './msgGroupVote';
+export * from './msgExecuteContract';
+export * from './msgInstantiateContract';
+export * from './msgInstantiateContract2';
+export * from './msgSend';
+export * from './msgSubmitGovProposal';
+export * from './msgSubmitGroupProposal';
+export * from './msgVoteWeighted';
+export * from './parameterChangeProposal';
+export * from './softwareUpgradeProposal';
+export * from './storeCodeProposal';
+export * from './textProposal';
