+++ src/services/messages/display/index.ts
@@ -1,0 +1,1 @@
+export * from './unpackDisplayObjectFromWalletMessage';
