+++ src/utils/display/index.ts
@@ -1,6 +1,7 @@
 export * from './capitalize';
 export * from './displayFormattedFields';
 export * from './displayMetadataFields';
+export * from './enumMapping';
 export * from './getJSType';
 export * from './hashFormat';
 export * from './numberFormat';
